% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Atys.R
\docType{class}
\name{Atys-class}
\alias{Atys-class}
\alias{Atys}
\title{SraRunInfo}
\description{
An S4-Object (summary object) that combines 3 class types
for the same dataset: Serratus class, SraRunInfo class, and Genome class
}
\section{Slots}{

\describe{
\item{\code{Serratus}}{S4 Serratus object}

\item{\code{SRA}}{S4 SraRunInfo object}

\item{\code{GENOME}}{S4 Genome object (character place holder for now)}
}}

\keyword{Genome,}
\keyword{Serratus,}
\keyword{SraRunInfo,}
\keyword{object}
