% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtilities.R
\name{checkNumericDifference}
\alias{checkNumericDifference}
\title{Check Numeric Difference}
\usage{
checkNumericDifference(x, output)
}
\arguments{
\item{x}{result object}
}
\value{
A boolean which is True if not all numerical value are present in both strings
}
\description{
Find and sort all the numeric values in the old and new value and returns if they are differences.
}
