% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compareVocabData.r
\name{compareVocabData}
\alias{compareVocabData}
\title{Compare two vocabularies}
\usage{
compareVocabData(connectionDetails, cdmDatabaseSchema,
  oldVocabularyDatabaseSchema = cdmDatabaseSchema,
  newVocabularyDatabaseSchema, findPrevalences = FALSE,
  oracleTempSchema = NULL)
}
\arguments{
\item{connectionDetails}{An R object of type\cr\code{connectionDetails} created using
the function \code{createConnectionDetails} in the
\code{DatabaseConnector} package.}

\item{cdmDatabaseSchema}{The name of the database schema that contains the OMOP CDM
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_instance.dbo'.}

\item{oldVocabularyDatabaseSchema}{The name of the database schema that contains the old
vocabulary instance.  Requires read permissions to this
database. On SQL Server, this should specifiy both the database
and the schema, so for example 'cdm_vocab.dbo'.}

\item{newVocabularyDatabaseSchema}{The name of the database schema that contains the ewvocabulary
instance.  Requires read permissions to this database. On SQL
Server, this should specifiy both the database and the schema,
so for example 'cdm_vocab.dbo'.}

\item{findPrevalences}{A boolean flag to indicate whether or not run
FetchPrevalences.sql}

\item{oracleTempSchema}{For Oracle only: the name of the database schema where you want
all temporary tables to be managed. Requires create/insert
permissions to this database.}
}
\description{
This function enables you to compare two vocabularies at the row level.
}
