% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractString.R
\name{extracSrting}
\alias{extracSrting}
\title{extract strings from the heading}
\usage{
extracSrting(inputString)
}
\arguments{
\item{inputString}{string with the value}
}
\value{
string value
}
\description{
For summary headers, extract string value
description=string;
}
