% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractNumber.R
\name{extractNumber}
\alias{extractNumber}
\title{extract numbers from the heading}
\usage{
extractNumber(inputString)
}
\arguments{
\item{inputString}{string with the value}
}
\value{
value
}
\description{
For summary headers, extract value
description=value;
}
