% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyPctID.R
\name{familyPctID}
\alias{familyPctID}
\title{Plot a histogram of percentage identity for a given family}
\usage{
familyPctID(
  family_df,
  family_name,
  score_threshold = 0,
  title = "",
  x1 = 0,
  x2 = 100,
  log = F
)
}
\arguments{
\item{family_df}{family data frame}

\item{family_name}{name of the family}

\item{score_threshold}{don't consider family hits with score below this value (0 by default)}

\item{x1}{left limit on x axis}

\item{x2}{right limit on x axis}

\item{log}{log10 y axis or not (default F)}
}
\value{
ggplot
}
\description{
Plot a histogram of percentage identity for a given family
}
\examples{

familyPctID(ZOO, "Coronaviridae", score_threshold = 50, title="", x1=75, x2=100)

}
