% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyScore.R
\name{familyScore}
\alias{familyScore}
\title{Plot a histogram of scores for a given family}
\usage{
familyScore(family_df, family_name, x1 = 0, x2 = 100, log = F)
}
\arguments{
\item{family_df}{family data frame}

\item{family_name}{name of the family}

\item{x1}{left limit on x axis}

\item{x2}{right limit on x axis}

\item{log}{log10 y axis or not (default F)}
}
\value{
ggplot
}
\description{
Plot a histogram of scores for a given family
}
\examples{

familyScore(ZOO, "Coronaviridae")

}
