% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBam.R
\name{getBAM}
\alias{getBAM}
\title{Downloads an SRA accession from an experiment S3 directory
1) download bam file
2) sort bam file (samtools)
3) index bam file (samtools)}
\usage{
getBAM(
  sra,
  s3path = "s3://serratus-public/out/200528_viro/bam/",
  local_path = "bam"
)
}
\arguments{
\item{sra}{SRA accession to download}

\item{local_path}{local path into which to download the data [bam]}

\item{s3_path}{a string s3 path holding .summary files}
}
\description{
Implicit system dependcies
- samtools
- aws cli
}
\keyword{S3,}
\keyword{Tantalus,}
\keyword{bam,}
\keyword{samtools}
