% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getHeatmap.R
\name{getHeatmap}
\alias{getHeatmap}
\title{getHeatmap}
\usage{
getHeatmap(accession, files)
}
\arguments{
\item{accession}{Accession, for which coverage should be plotted}

\item{files}{Files to parse}
}
\value{
plotly figure
}
\description{
Build a heatmap for the provided accession and SRA files
}
\keyword{Serratus,}
\keyword{accession,}
\keyword{coverage,}
\keyword{plotly}
