% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StringUtilities.R
\name{getNumericFromString}
\alias{getNumericFromString}
\title{Get Numerical Values from String}
\usage{
getNumericFromString(str)
}
\arguments{
\item{str}{String to check}
}
\value{
An ordered string list of numerical values
}
\description{
Retrieves and order list of numerical values in a string
}
