% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReference.R
\name{getReference}
\alias{getReference}
\title{Downlaods a Serratus Reference Sequence
and associated meta-data
* Fasta File
* Fasta Index
* Sumzer Index
* Masked regions bed file
* taxonomy file (Not Implemented)}
\usage{
getReference(
  refName = "cov3ma",
  s3_path = "s3://lovelywater/seq/",
  local_path = "seq"
)
}
\arguments{
\item{refName}{a string of the reference sequence base name}

\item{s3_path}{a string s3 path to the seq/ folder with references}

\item{local_path}{local path into which to download the data [seq]}
}
\description{
Implicit system dependcies
- aws cli
- gzip
}
\examples{

getReference(refName = "cov3ma",
             s3_path = "s3://serratus-public/seq/",
             local_path = "seq")

}
\keyword{S3,}
\keyword{Tantalus,}
\keyword{fasta,}
\keyword{genome,}
\keyword{meta-data}
\keyword{pan-genome,}
\keyword{reference,}
