% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummary.R
\name{getSummary}
\alias{getSummary}
\title{Downloads an S3 directory of Serratus .summary files
to local filesystem.
There is an implicit `aws cli` dependency here
Your system should have `aws s3` properly configured for this function
to work.}
\usage{
getSummary(s3_path, local_path)
}
\arguments{
\item{s3_path}{a string s3 path holding .summary files}

\item{local_path}{local path into which to download the data}
}
\description{
Downloads an S3 directory of Serratus .summary files
to local filesystem.
There is an implicit `aws cli` dependency here
Your system should have `aws s3` properly configured for this function
to work.
}
\examples{
getSummary( s3_path = 's3://serratus-public/out/200528_viro',
            local_path = '200528_viro')
}
\keyword{S3,}
\keyword{Tantalus,}
\keyword{summary}
