% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSummaryFiles.R
\name{getSummaryFiles}
\alias{getSummaryFiles}
\title{getSummaryFiles
Downloads a specified list of summary files and parses them
into a Serratus S4 object}
\usage{
getSummaryFiles(
  sras,
  s3_path = "s3://lovelywater2/summary/",
  local_path = "summary_files"
)
}
\arguments{
\item{sras}{a string vector of the SRAs to download}

\item{s3_path}{a string s3 path to summary files}

\item{local_path}{local path into which to download the data}
}
\value{
Serratus S4 class
}
\description{
Implicit system dependcies
- aws cli
}
\examples{
sras <- c("SRR9701190", "SRR9705127", "SRR975462", "SRR9843091", "SRR9843092")
serr <- getSummaryFiles(sras, 's3://lovelywater2/summary/', "summary_files")

}
\keyword{S4,}
\keyword{Serratus,}
\keyword{Tantalus,}
\keyword{sra}
