% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ShinyApps.r
\name{launchComparisonExplorer}
\alias{launchComparisonExplorer}
\title{Launch the comparison explorer}
\usage{
launchComparisonExplorer(results, launch.browser = TRUE)
}
\arguments{
\item{results}{The data object created by the \code{\link{compareVocabularies}}.}

\item{launch.browser}{Should the app be launched in your default browser (TRUE), or in a Shiny
window (FALSE)?}
}
\description{
Launch the comparison explorer
}
\details{
Launches a Shiny app that allows the user to explore the results of a vocabulary comparison.
}
