% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseAtys.R
\name{parseAtys}
\alias{parseAtys}
\title{Parse folder with Atys objects and retreive info about number of unique SRAs, GBs of data analyzed and sum of nucleotides}
\usage{
parseAtys(folder_with_atys)
}
\arguments{
\item{folder_with_atys}{folder with atys objects}
}
\value{
list with 3 values
}
\description{
Parse folder with Atys objects and retreive info about number of unique SRAs, GBs of data analyzed and sum of nucleotides
}
\examples{

parseAtys(DATA_PATH)

}
