% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseFamilies.R
\name{parseFamilies}
\alias{parseFamilies}
\title{Extract a family information from a file line}
\usage{
parseFamilies(hit, sra = "")
}
\arguments{
\item{hit}{string line with family accession}

\item{sra}{SRA identifier}
}
\value{
data frame (with one row for one file)
}
\description{
Extract a family information from a file line
}
