% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseHits.R
\name{parseHits}
\alias{parseHits}
\title{Extract an organism information from a file line}
\usage{
parseHits(hit, sra = "")
}
\arguments{
\item{hit}{string line with organism accession}

\item{sra}{SRA identifier}
}
\value{
data frame (with one row for one file)
}
\description{
Extract an organism information from a file line
}
