% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseSummaryFiles.R
\name{parseSummaryFiles}
\alias{parseSummaryFiles}
\title{Parse specified summary files into compressed fst data frames
which contain either family or accession info.}
\usage{
parseSummaryFiles(
  files,
  outputFolder = "parsed_df/",
  parseFamily = T,
  compressLevel = 100
)
}
\arguments{
\item{files}{The path to summary files}

\item{outputFolder}{The output folder}

\item{parseFamily}{True if you want to extract family information, False - if accessions}

\item{compressLevel}{Level of df compression (fst parameter), default is 100}
}
\value{
None
}
\description{
Parse specified summary files into compressed fst data frames
which contain either family or accession info.
}
\keyword{Serratus,}
\keyword{parsing}
\keyword{summary,}
