% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctidInScoreIntervals.R
\name{pctidInScoreIntervals}
\alias{pctidInScoreIntervals}
\title{Plot a histogram of pctid for the given family within family score intervals}
\usage{
pctidInScoreIntervals(
  family_df,
  family_name = NULL,
  title = "",
  scale_log = F,
  bin_scores = T,
  nucleotide = T,
  keep_levels = F
)
}
\arguments{
\item{family_df}{family data frame}

\item{family_name}{name of the family (single value)}

\item{title}{title of the plot}

\item{scale_log}{if plot log scale (not implemented)}

\item{bin_scores}{if split scores into bins for plotting}

\item{nucleotide}{TRUE/FALSE if nt or aa plot should be made [TRUE]}

\item{keep_levels}{TODO TRUE/FALSE should factor level of "family_name" be kept}
}
\value{
ggplot
}
\description{
Plot a histogram of pctid for the given family within family score intervals
}
\examples{

pctidInScoreIntervals(ZOO, "Coronaviridae", title="Coronaviridae in ZOO")

}
