% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDFs.R
\name{readDFs}
\alias{readDFs}
\title{readDFs}
\usage{
readDFs(files, all = T, columns = NULL, family = NULL, acc = NULL)
}
\arguments{
\item{files}{Files to parse into a single df}

\item{all}{DEPRECATED: If you want to parse all columns (True by default)}

\item{columns}{character vector of columns to return. NULL returns all [Default]}

\item{family}{character, if specified return only matching "family"}

\item{acc}{character, if specified return only matching virus accesion "top"}
}
\value{
data frame
}
\description{
Read fst serialized objects into data frame
}
\keyword{Serratus,}
\keyword{parsing}
\keyword{summary,}
