% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHeader.R
\name{readHeader}
\alias{readHeader}
\title{Extract header information}
\usage{
readHeader(sumFile)
}
\arguments{
\item{sumFile}{summary file}
}
\value{
header data frame (with one row for one file)
}
\description{
Extract header information
}
