% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHits.R
\name{readHits}
\alias{readHits}
\title{Extract all family or organism information from the provided file}
\usage{
readHits(sumFile, family = FALSE)
}
\arguments{
\item{sumFile}{Serratus summary file}

\item{family}{If True - extract family information; False - extract organism information (default)}
}
\value{
data frame with extracted information
}
\description{
Extract all family or organism information from the provided file
}
