% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSerratus.R
\name{readSerratus}
\alias{readSerratus}
\title{Parse list of files into S4 Serratus class}
\usage{
readSerratus(sumFiles)
}
\arguments{
\item{sumFiles}{Serratus summary file}
}
\value{
Serratus S4 class
}
\description{
Parse list of files into S4 Serratus class
}
