% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSra.R
\name{readSra}
\alias{readSra}
\title{read SRA identifiers from provided files}
\usage{
readSra(sumFile)
}
\arguments{
\item{sumFile}{the name of the summary file "SRA_ID.summary"}
}
\value{
SRA identifier
}
\description{
Creates a vector of SRA identifiers
}
