% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSraRunInfo.R
\name{readSraRunInfo}
\alias{readSraRunInfo}
\title{Parse a SraRunInfo csv file into a clean data.frame
and return a SraRunInfo S4 object}
\usage{
readSraRunInfo(sraCSV)
}
\arguments{
\item{sraCSV}{The path to an sraRunInfo.csv file}
}
\value{
SraRunInfo S4 object
}
\description{
Parse a SraRunInfo csv file into a clean data.frame
and return a SraRunInfo S4 object
}
\keyword{SRA,}
\keyword{Serratus,}
\keyword{SraRunInfo,}
\keyword{parsing}
