% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveResults.r
\name{saveComparisonResults}
\alias{saveComparisonResults}
\title{Save results to Shiny App}
\usage{
saveComparisonResults(results)
}
\arguments{
\item{results}{The data object created by the \code{\link{compareVocabularies}}.}
}
\description{
Save results to Shiny App
}
\details{
Saves results out to a data frame for use in a published shiny app.
}
