% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saveResults.r
\name{saveDiffResults}
\alias{saveDiffResults}
\title{Save json diff Results to Shiny App}
\usage{
saveDiffResults(diffResults)
}
\arguments{
\item{jsonResults}{The data object created by the \code{\link{createDiffSummary}}.}
}
\description{
Save json diff Results to Shiny App
}
\details{
Saves results out to json file for use in a published shiny app.
}
