% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SeqReference.R
\docType{class}
\name{seqReference-class}
\alias{seqReference-class}
\alias{seqReference}
\title{seqReference}
\description{
An S4-Object definition for the Pan-Genome / Query Reference
}
\details{
Contains:
 - Fasta index of genome (.fa.fai)
 - Summarizer Meta-data file (.sumzer.tsv)
 - Serratax taxonomy information (.serratax) [not implemented]
}
\section{Slots}{

\describe{
\item{\code{rname}}{String of reference version name}

\item{\code{fai}}{Data.frame for a fasta index file}

\item{\code{sumzer}}{Data.frame of the summarizer meta-data}

\item{\code{tax}}{Data.frame of the serratax output taxonomy information}
}}

\keyword{Serratus,}
\keyword{genome,}
\keyword{meta-data}
\keyword{object,}
\keyword{pan-genome,}
\keyword{reference,}
\keyword{summary,}
