% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Serratus.R
\docType{class}
\name{Serratus-class}
\alias{Serratus-class}
\alias{Serratus}
\title{An S4 class to store parsed summary files.}
\description{
An S4 class to store parsed summary files.
}
\section{Slots}{

\describe{
\item{\code{sra}}{A vector of SRA identifiers (one identifier per file)}

\item{\code{header}}{A data frame with header information for each file (genome, date)}

\item{\code{family}}{A data frame with information about detected families}

\item{\code{acc}}{A data frame with information about detected organisms}
}}

