% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SraRunInfo.R
\docType{class}
\name{SraRunInfo-class}
\alias{SraRunInfo-class}
\alias{SraRunInfo}
\title{SraRunInfo}
\description{
An S4-Object definition for the SRA's "runInfo"
csv files. Downloaded from website.
Send to --> File --> runInfo
}
\section{Slots}{

\describe{
\item{\code{RunInfo}}{Data.frame of SRA "runInfo" csv file}
}}

\keyword{SRA}
\keyword{Serratus,}
\keyword{object,}
\keyword{runInfo,}
