% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Tantalus.R
\docType{package}
\name{Tantalus}
\alias{Tantalus}
\title{Tantalus package}
\description{
R Data Exploration for Serratus
}
\author{
Gherman Novakovsky \email{gnovakovsky@cmmt.ubc.ca}, Artem Babaian \email{ababaian@bccrc.ca}
}
