% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyScore.R
\name{familyScore}
\alias{familyScore}
\title{Plot a histogram of scores for a given family}
\usage{
familyScore(serratus, family_name, x1 = 0, x2 = 100)
}
\arguments{
\item{serratus}{serattus S4 object with family data frame}

\item{family_name}{name of the family}

\item{x1}{left limit on x axis}

\item{x2}{right limit on x axis}
}
\value{
ggplot
}
\description{
Plot a histogram of scores for a given family
}
\examples{

familyScore(ZOO, "Coronaviridae")

}
