% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/familyScorPctID.R
\name{familyScorPctID}
\alias{familyScorPctID}
\title{Plot a scatterplot/histogram of percentage identity vs family score for a given family}
\usage{
familyScorPctID(serratus, family_name)
}
\arguments{
\item{serratus}{serattus S4 object with family data frame}

\item{family_name}{name of the family}
}
\value{
ggplot
}
\description{
Plot a scatterplot/histogram of percentage identity vs family score for a given family
}
\examples{

familyScorPctID(ZOO, "Coronaviridae")

}
