% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getBam.R
\name{getBAM}
\alias{getBAM}
\title{Downloads an SRA accession from an experiment S3 directory
1) download bam file
2) sort bam file (samtools)
3) index bam file (samtools)}
\usage{
getBAM(sra, s3_path = "s3://lovelywater/", local_path = "bam")
}
\arguments{
\item{sra}{SRA accession to download}

\item{s3_path}{a string s3 path holding .summary files}

\item{local_path}{local path into which to download the data [bam]}
}
\description{
Implicit system dependcies
- samtools
- aws cli
}
\examples{

getBam(sra = 'SRR5447135', s3_path = 's3://lovelywater/bam/')

}
\keyword{S3,}
\keyword{Tantalus,}
\keyword{bam,}
\keyword{samtools}
