% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getLoadedSummaryFiles.R
\name{getLoadedSummaryFiles}
\alias{getLoadedSummaryFiles}
\title{getLoadedSummaryFiles
Parses the specified list of SRA accessions (which are supposed to be present in the specified folder)
into a Serratus S4 object}
\usage{
getLoadedSummaryFiles(sras, path_to_files = "")
}
\arguments{
\item{sras}{a string vector of the SRAs}

\item{path_to_files}{a local folder with the SRA summary files}
}
\value{
Serratus S4 class
}
\description{
Implicit system dependcies
- aws cli
}
\examples{
sras <- c("SRR9701190", "SRR9705127", "SRR975462", "SRR9843091", "SRR9843092")
serr <- getSummaryFiles(sras, 's3://serratus-public/out/200528_viro/summary/', "summary_files")

}
