% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pctidInScoreIntervals.R
\name{pctidInScoreIntervals}
\alias{pctidInScoreIntervals}
\title{Plot a histogram of pctid for the given family within family score intervals}
\usage{
pctidInScoreIntervals(
  family_df,
  family_name = NULL,
  title = "",
  scale_log = F,
  bin_scores = T
)
}
\arguments{
\item{family_df}{family data frame}

\item{family_name}{name of the family}

\item{title}{title of the plot}
}
\value{
ggplot
}
\description{
Plot a histogram of pctid for the given family within family score intervals
}
\examples{

pctidInScoreIntervals(ZOO, "Coronaviridae", title="Coronaviridae in ZOO")

}
