% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDFs.R
\name{readDFs}
\alias{readDFs}
\title{readDFs}
\usage{
readDFs(files, all = T, columns = c())
}
\arguments{
\item{files}{Files to parse into a single df}

\item{all}{If you want to parse all columns (True by default)}

\item{columns}{if all is False, you need to provide a vector of column names you want to extract}
}
\value{
data frame
}
\description{
Read fst serialized objects into data frame
}
\keyword{Serratus,}
\keyword{parsing}
\keyword{summary,}
