% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDfSQL.R
\name{readDfSQL}
\alias{readDfSQL}
\title{readDfSQL}
\usage{
readDfSQL(
  dbs,
  table,
  columns = NULL,
  family = NULL,
  acc = NULL,
  sras = NULL,
  dataframe = F
)
}
\arguments{
\item{dbs}{Database object (SQL)}

\item{table}{Table in the database}

\item{columns}{character vector of columns to return. NULL returns all [Default]}

\item{family}{character, if specified return only matching "Family"}

\item{acc}{character, if specified return only matching virus accesion "Top"}

\item{sras}{character, return only specified accessions}

\item{dataframe}{boolean, if T, convert the output to dataframe}
}
\value{
data frame
}
\description{
Read fst serialized objects into data frame
}
\keyword{Serratus,}
\keyword{parsing}
\keyword{summary,}
