% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scoreInPctidIntervals.R
\name{scoreInPctidIntervals}
\alias{scoreInPctidIntervals}
\title{Plot a histogram of family score for the given family within pctid score intervals}
\usage{
scoreInPctidIntervals(family_df, family_name, title = "")
}
\arguments{
\item{family_df}{family data frame}

\item{family_name}{name of the family}

\item{title}{title of the plot}
}
\value{
ggplot
}
\description{
Plot a histogram of family score for the given family within pctid score intervals
}
\examples{

scoreInPctidIntervals(ZOO, "Coronaviridae", title="Coronaviridae in ZOO")

}
